Adding in code for project 13
Code entails sensing the voltage of the TX controller from the RX-Receiver Side which anticipates a voltage current coming in when the LED is on and 0V when the LED is off. In the same code, we'll utilize the FFT library to identify the frequency of the LED from the RX-receiver.